requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
python-dotenv
jinja2
python-multipart
edge-tts
pandas
openpyxl
python-docx
python-pptx
